Move ThroughputThrottler class from clients to server-common
